requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
strands-agents[ollama]>=0.1.0
pydantic>=2.0.0
ollama>=0.1.6
jinja2